First pass at a conditional wizard panel for no_sandboxes experiment
This a general pattern for using the course object in conjunction with the panel properties to selectively include a panel only of the course has a corresponding property. This way, we can give returning instructors the option to go the no_sandboxes route, while ensuring that first-time instructors get randomly assigned as part of the experiment.