Make WizardTimelineManager handle no_sandboxes tags and flags
I'm taking a moment to appreciate how I implemented the handling of tags and flags in this code. I started writing special handling for dealing with setting and unsetting the flag and tag for 'no_sandboxes' so that it could be undone with a second pass through the wizard, and realized that all the necessary steps are already taken via the flag and tag handling we already have.